Refactor: rename into `HydraAuctionUtils.WebSockets.*`
